Initial commit: LICENSE, README, and .gitignore
I've gone with a Creative Commons Attribution+ShareAlike+NonCommercial
license, as I want people to be able to do stuff with my music, except
of course sell it without my permission.

I've also detailed a bit of the story in how these pieces came to be,
cause people care about that sort of thing.